Port "move pixelate filter to gd_filter.c" (sync with upstream)
Cf. <https://github.com/libgd/libgd/commit/0f0894a7dd5c74cb9824f5c297569434c0f5b12e>.
