+++ ext/gd/config.m4
@@ -233,7 +233,6 @@ if test "$PHP_GD" != "no"; then
       libgd/gd_io.c
       libgd/gd_jpeg.c
       libgd/gd_matrix.c
-      libgd/gd_pixelate.c
       libgd/gd_png.c
       libgd/gd_rotate.c
       libgd/gd_security.c
